agregando alertas a accion

+++ src/app/ComponentesVista/Mantenimiento/acciones/acciones.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MantAcciones } from 'src/app/Modelo/MantAcciones';
 import { MantenimientoService } from 'src/app/service/mantenimiento.service';
 import { Router, ActivatedRoute } from '@angular/router';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-acciones',
@@ -43,12 +44,36 @@ export class AccionesComponent implements OnInit {
       console.log(accion);
       this.cargarAcciones();
     });
+    Swal.fire({
+      position: 'top-end',
+      icon: 'success',
+      title: 'Se ha actualizado con exito!',
+      showConfirmButton: false,
+      timer: 2500
+    });
   }
 
   eliminar(accion:MantAcciones):void{
-    this.accionesService.eliminarAcciones(accion.idmantacciones).subscribe(response=>{
-      this.requests = this.requests.filter(req=>req!==accion)
-    });
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Esta accion no se podra revertir!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        this.accionesService.eliminarAcciones(accion.idmantacciones).subscribe(response=>{
+          this.requests = this.requests.filter(req=>req!==accion)
+        });
+        Swal.fire(
+          'Eliminado!',
+          'La accion ha sido eliminada',
+          'success'
+        )
+      }
+    })
   }
 
   public crearAccion(): void{
@@ -56,6 +81,13 @@ export class AccionesComponent implements OnInit {
     this.accionesService.createAcciones(this.mantAcciones).subscribe(data => {
       this.router.navigate([`home/acciones/`+this.tipo]);
       this.cargarAcciones();
+    });
+    Swal.fire({
+      position: 'top-end',
+      icon: 'success',
+      title: 'Se ha creado con exito!',
+      showConfirmButton: false,
+      timer: 2500
     });
     this.mantAcciones.nombre="";
     this.mantAcciones.tipo=null;
